lts-internal/philosophy-updater/requirements.txt
google-generativeai>=0.8.0
anthropic>=0.40.0
python-dotenv>=1.0.0
watchdog>=3.0.0
requests>=2.31.0
